backend/requirements.txt
# Flask
# flask-cors
# pymongo
# openai-whisper
# opencv-python
# Pillow
# nltk
# librosa
# torch
# torchaudio
# requests
# googletrans

Flask
flask-cors
flask-bcrypt
flask-jwt-extended
pymongo
openai-whisper
opencv-python
Pillow
nltk
librosa
torch
torchaudio
requests
googletrans
python-dotenv
bcrypt